+++ R/options.R
@@ -3,6 +3,18 @@ options::define_options(
   "tty refresh interval when reporting results in milliseconds",
   tty_tick_interval = 0.1,
 
+  "deafult tty height used for the ANSI reporter. Used only
+   if correct values could not be acquired with system('tput lines')",
+  tty_default_height = 50,
+
+  "`logical`, indicating whether additional garbage collection should be
+   performed before starting a new task, if at least one process recently
+   finalized. This can cause the checker to orchestrate tasks slower but
+   is recommended to be used for designs with many sub-processes required as
+   native garbage collection can lag leading to memory issues. Disable only
+   when maximum prefromance is required and memory is not the issue.",
+  proactive_gc = TRUE,
+
   "character vector indicating whether R error should be thrown when issues
    are discovered when generating results. \"never\" means that no errors
    are thrown. If \"issues\" then errors are emitted only on issues, whereas